server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
passlib[bcrypt]
pyjwt
python-jose
